backend/requirements.txt
pymammotion>=0.1.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
websockets>=13.0
python-dotenv>=1.0.0
cryptography>=43.0.0
